checkAll-->check, check-->query
