+++ core/__tests__/tasks/appRefreshQuery/check.ts
@@ -5,23 +5,17 @@ import {
   Schedule,
   App,
   Run,
-  plugin,
   GrouparooModel,
   AppRefreshQuery,
 } from "../../../src";
-import { RunOps } from "../../../src/modules/ops/runs";
 
-describe("tasks/appRefreshQuery:check", () => {
-  let model: GrouparooModel;
-  let source: Source;
-  let schedule: Schedule;
-
+describe("tasks/appRefreshQueries:check", () => {
   helper.grouparooTestServer({ truncate: true, enableTestPlugin: true });
   beforeEach(async () => await api.resque.queue.connection.redis.flushdb());
 
   beforeAll(async () => {});
 
-  describe("appRefreshQuery:run", () => {
+  describe("appRefreshQueries:check", () => {
     let model: GrouparooModel;
     let app: App;
     let source: Source;
@@ -48,37 +42,31 @@ describe("tasks/appRefreshQuery:check", () => {
     });
 
     test("can be enqueued and run", async () => {
-      await task.enqueue("appRefreshQuery:check", { appRefreshQuery }); //does not throw
+      await task.enqueue("appRefreshQueries:check", {}); //does not throw
 
-      await specHelper.runTask("appRefreshQuery:check", { appRefreshQuery });
+      await specHelper.runTask("appRefreshQueries:check", {});
 
-      const enqueuedRuns = await Run.findAll({
-        where: { creatorId: schedule.id, state: "running" },
-      });
-      expect(enqueuedRuns.length).toBe(1);
-
-      Run.truncate();
+      const found = await specHelper.findEnqueuedTasks(
+        "appRefreshQueries:check"
+      );
+      expect(found.length).toEqual(1);
     });
 
-    test("updates value and timestamps if value is updated", async () => {
-      await task.enqueue("appRefreshQuery:check", { appRefreshQuery }); //does not throw
+    test("enqueues appRefreshQuery:query", async () => {
+      await task.enqueue("appRefreshQueries:check", {}); //does not throw
 
-      await specHelper.runTask("appRefreshQuery:check", { appRefreshQuery });
+      await specHelper.runTask("appRefreshQueries:check", {});
 
-      appRefreshQuery.reload();
-
-      //test plugin always returns a timestamp
-      expect(appRefreshQuery.value.length).toBe(13);
-      expect(appRefreshQuery.lastChangedAt).toBeTruthy();
-      expect(appRefreshQuery.lastConfirmedAt).toBeTruthy();
-
-      Run.truncate();
+      const foundChildren = await specHelper.findEnqueuedTasks(
+        "appRefreshQuery:query"
+      );
+      expect(foundChildren.length).toEqual(1);
     });
 
     test("does not throw if no appRefreshQueries found", async () => {
       appRefreshQuery.destroy();
 
-      await task.enqueue("appRefreshQuery:check", { appRefreshQuery }); //does not throw
+      await task.enqueue("appRefreshQueries:check", {}); //does not throw
 
       //also does not enqueue any runs
       const enqueuedRuns = await Run.findAll({
@@ -87,35 +75,23 @@ describe("tasks/appRefreshQuery:check", () => {
       expect(enqueuedRuns.length).toBe(0);
     });
 
-    test("does not throw if no app or schedules ready for appRefreshQuery", async () => {
+    test("does not throw if no app or schedules ready for appRefreshQueries", async () => {
       const anotherApp = await helper.factories.app();
       await anotherApp.update({ state: "deleted" });
 
       new AppRefreshQuery({
-        appId: anotherApp.id,
+        appId: app.id,
         refreshQuery: "SELECT 'hi' AS name;",
         state: "ready",
       });
       await appRefreshQuery.save();
 
-      await task.enqueue("appRefreshQuery:check", { appRefreshQuery }); //does not throw
+      await task.enqueue("appRefreshQueries:check", {}); //does not throw
 
       //also does not enqueue any runs
       const enqueuedRuns = await Run.findAll({
         where: { creatorId: schedule.id, state: "running" },
       });
-      expect(enqueuedRuns.length).toBe(0);
-    });
-
-    test("only schedules marked 'refreshEnabled:true' are triggered", async () => {
-      await schedule.update({ refreshEnabled: false });
-
-      await task.enqueue("appRefreshQuery:check", { appRefreshQuery }); //does not throw
-
-      const enqueuedRuns = await Run.findAll({
-        where: { creatorId: schedule.id, state: "running" },
-      });
-
       expect(enqueuedRuns.length).toBe(0);
     });
   });
